Integrate controllers and routes files with old structure, set up structure for edit inventory item api, in progress
