+++ server.js
@@ -3,7 +3,8 @@ const app = express();
 require("dotenv").config();
 app.use(express.json());
 const cors = require('cors')
-const warehouseRoutes = require('./routes/routes.js');
+const warehouseRoutes = require('./routes/routes-warehouse.js');
+//inventory route import like this one ^
 
 
 
@@ -14,9 +15,10 @@ const PORT = process.env.PORT || 5000;
 const CORS_ORIGIN = process.env.CORS_ORIGIN || "http://localhost:3000";
 app.use(cors({ origin: CORS_ORIGIN }));
 
+app.use('/api/warehouses/', warehouseRoutes);
+// app.use('/api', warehouseRoutes); 
 
-app.use('/api', warehouseRoutes); 
-
+//app.use('/api/inventories/', inventoryRoutes);
 
 app.listen(PORT, () =>{
     console.log(`listening on port ${PORT}`);
